moved formatDate and formatPhoneNumber functions from hocs to lib, removed hocs folder, updated imports and implementations accordingly, and refactored filterFields interfaces and implementations in adminRegistrationApplications and emergencyEvacuationApplications modules.
