--- src/app/hocs/formatPhoneNumber.ts
@@ -1,4 +1,0 @@
-export const formatPhoneNumber = (countryCode: string, lineNumber: string): string => {
-    const formattedLineNumber = lineNumber.replace(/(\d{3})(\d{3})(\d{2})(\d{2})/, '($1) $2 $3 $4')
-    return `+${countryCode} ${formattedLineNumber}`
-  }
