+++ src/modules/adminRegistrationApplications/constants/filterFields.ts
@@ -1,16 +1,6 @@
 import { StatusData } from './status'
 
-interface FilterFields {
-  label: string
-  value: string
-  type?: string
-  placeholder?: string
-  options?: { label: string; value: string }[]
-  maxLength?: number
-  fieldsType?: string
-}
-
-const filterFields: FilterFields[] = [
+const filterFields: FilterFields = [
   {
     label: 'status',
     value: 'status',
