+++ src/types/index.d.ts
@@ -1,0 +1,9 @@
+declare type FilterFields = {
+  label: string
+  value: string
+  type?: string
+  placeholder?: string
+  options?: { label: string; value: string }[]
+  maxLength?: number
+  fieldsType?: 'selectBoxField' | 'inputField' | 'quickFilterField'
+}[]
